Add typescript declarations
