+++ package.json
@@ -4,10 +4,12 @@
   "description": "React-fiber renderer for zdog",
   "main": "dist/index.cjs.js",
   "module": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "typings": "dist/index.d.ts",
   "sideEffects": false,
   "scripts": {
     "prebuild": "rimraf dist",
-    "build": "rollup -c",
+    "build": "rollup -c rollup.config.ts && tsc",
     "prepare": "npm run build",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -54,7 +56,6 @@
     "zdog": ">=1.1"
   },
   "devDependencies": {
-    "zdog": "^1.1.0",
     "@babel/core": "7.4.4",
     "@babel/plugin-proposal-class-properties": "7.4.4",
     "@babel/plugin-proposal-do-expressions": "7.2.0",
@@ -68,6 +69,7 @@
     "@babel/preset-typescript": "^7.3.3",
     "@types/lodash-es": "^4.17.3",
     "@types/react": "^16.8.15",
+    "@types/zdog": "^1.1.1",
     "babel-eslint": "^10.0.1",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "husky": "^2.1.0",
@@ -79,6 +81,8 @@
     "rollup-plugin-babel": "^4.3.2",
     "rollup-plugin-commonjs": "^9.3.4",
     "rollup-plugin-node-resolve": "^4.2.3",
-    "rollup-plugin-size-snapshot": "^0.8.0"
+    "rollup-plugin-size-snapshot": "^0.8.0",
+    "typescript": "^3.9.5",
+    "zdog": "^1.1.0"
   }
 }
